Adds Gyro support to the simulator, and Unit tests for the gyros
As I asked in #77, this has some hacks implemented for gyro simulation.
It was going to be next to impossible to simulate AnalogGyro properly,
and ADXRS450 would have been easier, but still much more difficult to
implement.

+++ WPILib/AnalogInput.cs
@@ -260,10 +260,10 @@ namespace WPILib
         /// <summary>
         /// Read the number of accumulated values
         /// </summary>
-        public long GetAccumulatorCount()
+        public uint GetAccumulatorCount()
         {
             int status = 0;
-            long value = HALAnalog.GetAccumulatorCount(m_port, ref status);
+            uint value = HALAnalog.GetAccumulatorCount(m_port, ref status);
             CheckStatus(status);
             return value;
         }
